Make the fluent tests more fluent.
No longer need to create Process or Task instances as a test author.
Methods return a ProcessTest or TaskTest instance and methods that
make sense on a process or on a task are moved into these classes.

Honestly I've left things pretty messy, mostly was hoping to make it
clear how we can further streamline the tests, and also make it so
that you only get methods that make sense when autocompleting on the
fluent chain...